% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarization.R
\name{binarize_exp}
\alias{binarize_exp}
\title{Binarize gene expression}
\usage{
binarize_exp(sce, binarize_cutoff = 0.2)
}
\arguments{
\item{sce}{SingleCellExperiment}
}
\value{

}
\description{
This function generates on/off binarized data for gene expression
}
