% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_TI.R
\name{convert_monocle2}
\alias{convert_monocle2}
\title{Convert monocle2 output into GeneSwitches object}
\usage{
convert_monocle2(monocle2_obj, states, expdata)
}
\arguments{
\item{monocle2_obj}{monocle2 output object}

\item{states}{a vector of states (path) that are interested in}

\item{logexpdata}{log-normal gene expression}
}
\value{
A SingleCellExperiment object
}
\description{
This function converts monocle2 output into GeneSwitches object
}
