% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_TI.R
\name{convert_slingshot}
\alias{convert_slingshot}
\title{Convert slingshot output into GeneSwitches object}
\usage{
convert_slingshot(sce_slingshot, pseudotime_idx, assayname = "expdata")
}
\arguments{
\item{sce_slingshot}{slingshot SingleCellExperiment output object}

\item{pseudotime_idx}{name of desired pseudotime path to apply GeneSwitches}

\item{assayname}{expression assay to use}
}
\value{

}
\description{
This function converts slingshot output into GeneSwitches object
}
