% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trajectories.R
\name{distinct_genes}
\alias{distinct_genes}
\title{Identify distinct switching genes for each path}
\usage{
distinct_genes(toplotgl_Rsub1, toplotgl_Rsub2, path1name = "Path1Genes",
  path2name = "Path2Genes", r2cutoff = 0.05, scale_timeline = FALSE,
  path1time = NULL, path2time = NULL, bin = 100)
}
\arguments{
\item{toplotgl_Rsub1}{switching genes of path1}

\item{toplotgl_Rsub2}{switching genes of path2}

\item{path1name}{name of path1 given by user}

\item{path2name}{name of path2 given by user}

\item{r2cutoff}{pseudo R^2 cutoff}
}
\value{

}
\description{
This function identifies distinct switching genes for each path
}
