% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{filter_switchgenes}
\alias{filter_switchgenes}
\title{Extract switching gene list of interesting}
\usage{
filter_switchgenes(
  sce,
  allgenes = FALSE,
  pathway_name = NULL,
  genelists = GeneSwitches:::gs_genelists,
  genetype = c("Surface proteins", "TFs"),
  zero_pct = 0.9,
  r2cutoff = 0.03,
  direction = c("up", "down"),
  topnum = 1e+05
)
}
\arguments{
\item{sce}{switching genes}

\item{allgenes}{if use all genes}

\item{pathway_name}{a list of pathway name(s) to plot}

\item{genelists}{a gene list to plot}

\item{genetype}{specific gene type to plot c("EMT", "reprogramming", "stem", "surface", "TF")}

\item{zero_pct}{zero-expression percentage cut off for significant genes}

\item{r2cutoff}{pseudo R^2 cutoff}

\item{direction}{switching direction, up or down}

\item{topnum}{number of top genes ordered by pseudo R^2 value}
}
\value{
A dataframe of filtered switching genes
}
\description{
This function extract a list of significant switching genes
}
