% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_regression.R
\name{find_switch_logistic_fastglm}
\alias{find_switch_logistic_fastglm}
\title{Fit fast logistic regression and find switching timepoint}
\usage{
find_switch_logistic_fastglm(
  sce,
  downsample = TRUE,
  ds_cutoff = 0.7,
  zero_ratio = ds_cutoff,
  sig_FDR = 0.05,
  ncores = 3,
  EPV = 10
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{downsample}{Logical. if do random downsampling of zeros}

\item{ds_cutoff}{only do downsampling if zero percentage is over this cutoff}

\item{zero_ratio}{downsampling zeros to this proportion}

\item{sig_FDR}{FDR cut off for significant genes}

\item{ncores}{Numeric. Number of cores to use for parallel processing. Default is 3.}

\item{EPV}{Numeric. Minimum number of Events Per Variable (EPV) required for a gene to be included in the regression analysis. Default is 10.}
}
\value{
A SingleCellExperiment object with the regression results added to the rowData.
}
\description{
This function fits fast logistic regression and find switching timepoint for each gene
}
