% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_switch_pathway.R
\name{find_switch_pathway}
\alias{find_switch_pathway}
\title{Find significantly changed pathways and switching timepoint}
\usage{
find_switch_pathway(scerowdata, pathways = msigdb_h_c2_c5, toplotgl_sig,
  sig_FDR = 0.05)
}
\arguments{
\item{pathways}{a list of pathways with genes}

\item{sig_FDR}{FDR cut off for significant pathways}

\item{sce}{SingleCellExperiment}

\item{toplotgl_ptw}{swiching genes to plot}
}
\value{

}
\description{
This function finds significantly changed pathways and determine
    the switching timepoint for each pathway
}
