% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_inputData.R
\name{get_example_inputData}
\alias{get_example_inputData}
\title{Download example input data}
\usage{
get_example_inputData()
}
\value{

}
\description{
This function checks if example input files are in the current directory and if not download them
}
