% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs_genelists}
\alias{gs_genelists}
\title{An example of gene lists}
\format{A data frame with 3071 rows and 2 columns. rows are genes, and columns are "genenames" and "genetypes".}
\source{
\url{http://wlab.ethz.ch/cspa/, http://humantfs.ccbr.utoronto.ca/}
}
\usage{
gs_genelists
}
\description{
Provided gene lists with surface proteins and transctription factors
}
\keyword{datasets}
