% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_h_c2_c5}
\alias{msigdb_h_c2_c5}
\title{An example of gene sets}
\format{A list with 6153 pathways, and genes in each pathway}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msigdb_h_c2_c5
}
\description{
Pathways of GSEA MSigDB hallmark, c2 and c5 gene sets
}
\keyword{datasets}
