% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_switch_pathway.R
\name{phyper_pathway}
\alias{phyper_pathway}
\title{Find significantly changed pathways and switching timepoint}
\usage{
phyper_pathway(N, pathways, toplotgl_ptw, sig_FDR = 0.05,
  direction = c("up", "down"))
}
\arguments{
\item{N}{expressed genes in pathways}

\item{pathways}{a list of pathways with genes}

\item{toplotgl_ptw}{swiching genes to plot}

\item{sig_FDR}{FDR cut off for significant pathways}

\item{direction}{switching direction, up or down}
}
\value{

}
\description{
Apply hypergeometric test to determine significantly changed
    pathways and determine the switching timepoint for each pathway
}
