% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_TI.R
\name{plot_monocle_State}
\alias{plot_monocle_State}
\title{plot monocle2 trajectory colored by State}
\usage{
plot_monocle_State(monocle2_obj)
}
\arguments{
\item{monocle2_obj}{monocle2 output object}
}
\value{
A grid object composed of two ggplot objects
}
\description{
This function plots monocle2 trajectory with "State" colors
}
