% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline_ggplot}
\alias{plot_timeline_ggplot}
\title{Plot switching genes}
\usage{
plot_timeline_ggplot(tml, timedata, iffulltml = TRUE, txtsize = 3.5,
  color_by = "feature_type")
}
\arguments{
\item{tml}{switching genes}

\item{timedata}{pseudotime for cells}

\item{iffulltml}{if plot the full timeline}

\item{txtsize}{text size for gene names}

\item{color_by}{the cell attribute (e.g. the column of tml) to map to each cell's color}
}
\value{

}
\description{
This function plots switching genes on the pseudo-timeline
}
