% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_switch_pathway.R
\name{reduce_pathways}
\alias{reduce_pathways}
\title{Reduce redundant pathways}
\usage{
reduce_pathways(switch_pw, pathways, redundant_pw_rate = 0.8)
}
\arguments{
\item{switch_pw}{a data frame with significantly changed pathways}

\item{pathways}{a list of pathways with genes}

\item{redundant_pw_rate}{cutoff ratio for merging redundant pathways}
}
\value{

}
\description{
Reduce significant pathways that are with same genes over certain ratio,
    do up- and down- regulated pathways separately
}
