% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_TI.R
\name{subset_pseudotime}
\alias{subset_pseudotime}
\title{Subset GeneSwitches object based on the range of pseudotime}
\usage{
subset_pseudotime(
  sce,
  min_time,
  max_time,
  assayname = "expdata",
  minexp = 0,
  mincells = 10
)
}
\arguments{
\item{sce}{GeneSwitches object which is a SingleCellExperiment object}

\item{min_time}{lower bound of pseudotime}

\item{max_time}{upper bound of pseudotime}

\item{assayname}{expression assay to use}

\item{minexp}{minimun expression to filer genes}

\item{mincells}{minimun cells with expression}
}
\value{
A subsetted SingleCellExperiment object
}
\description{
This function subsets GeneSwitches object based on the range of pseudotime
}
