% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarization.R
\name{binarize_exp}
\alias{binarize_exp}
\title{Binarize gene expression}
\usage{
binarize_exp(sce, fix_cutoff = FALSE, binarize_cutoff = 0.2,
  ncores = 3)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{fix_cutoff}{Logical. if use fixed global cutoff for binarization, default FALSE}

\item{binarize_cutoff}{fixed global cutoff for binarization, default 0.2}

\item{ncores}{number of cores}
}
\value{

}
\description{
This function generates on/off binarized data for gene expression
}
