% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trajectories.R
\name{common_genes_plot}
\alias{common_genes_plot}
\title{Plot common switching genes between paths}
\usage{
common_genes_plot(ggData, timedata)
}
\arguments{
\item{ggData}{data frame for common genes}

\item{timedata}{timedata to show on plot}
}
\value{
A ggplot object representing the common genes plot
}
\description{
This function plots common switching genes between two paths
}
