% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_exp.R
\name{plot_gene_exp}
\alias{plot_gene_exp}
\title{Plot gene expression}
\usage{
plot_gene_exp(
  sce,
  gene,
  reduction,
  fitting = FALSE,
  bin_width = 0.1,
  downsample = FALSE,
  ds_cutoff = 0.7,
  zero_ratio = 0.7,
  ptsize = 0.7
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{gene}{one gene of interest}

\item{reduction}{dimensional reduction method}

\item{fitting}{if plot logistic regression fitting}

\item{downsample}{if do random downsampling of zeros}

\item{ds_cutoff}{only do downsampling if zero percentage is over this cutoff}

\item{zero_ratio}{downsampling zeros to this proportion}

\item{ptsize}{point size}
}
\value{
A grid object or a ggplot object showing gene expression
}
\description{
This function plot gene expression on two-dimensional space
}
